backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "python-multipart>=0.0.7",
    "email-validator>=2.1.0.post1",
    "passlib[bcrypt]>=1.7.4",
    "tenacity>=8.2.3",
    "pydantic>=2.0.0",
    "emails>=0.6",
    "jinja2>=3.1.4",
    "alembic>=1.12.1",
    "httpx>=0.25.1",
    "psycopg[binary]>=3.1.13",
    "sqlmodel>=0.0.21",
    # Pin bcrypt to 3.2.2 to avoid passlib incompatibilty
    "bcrypt==3.2.2",
    "pydantic-settings>=2.2.1",
    "sentry-sdk[fastapi]>=1.40.6",
    "pyjwt>=2.8.0",
    "apscheduler>=3.10.0",
]

[dependency-groups]
dev = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
    "flake8<8.0.0,>=7.0.0",
    "flake8-bugbear",
    "flake8-comprehensions",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG001"]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
